+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-05-22 12:34:14
-// Commit: feat(ui): add async ui stream provider
+// Updated at: 2026-05-22 13:11:52
+// Commit: test(ui): implement widget integration tests for ui
 
 class uiComponentState {
-  final String commitRef = '1316';
-  final String lastMessage = 'feat(ui): add async ui stream provider';
-  final DateTime timestamp = DateTime.parse('2026-05-22T12:34:14+05:00');
+  final String commitRef = '1317';
+  final String lastMessage = 'test(ui): implement widget integration tests for ui';
+  final DateTime timestamp = DateTime.parse('2026-05-22T13:11:52+05:00');
 }